requirements.txt
pandas
numpy
requests
python-dotenv
websockets
alpaca-trade-api
pyarrow